Remove 1.8-1.9 upgrade codes of kubeadm
